server/memoryquest_server/requirements.txt
agent-framework --pre
fastapi[standard]
agent-framework-mem0 --pre
mem0ai
python-dotenv
cognee
cognee-community-vector-adapter-qdrant
azure-cli
hindsight-client